arch2-2pc-implementation/storage/requirements.txt
Flask==3.0.0
requests==2.31.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1